Extracted test to a separate file dual-output-end-stream.ts
